+++ CHANGES.rst
@@ -1,7 +1,8 @@
 0.8.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Base the cache key on the globs in Docker COPY/ADD commands.
+  (`Issues #98 <https://github.com/6si/shipwright/pull/98>`_).
 
 
 0.8.0 (2017-06-08)